Respond to reviewer comments.
