+++ extensions/dicom-sr/src/index.js
@@ -25,6 +25,7 @@ export default {
   id,
   dependencies: [
     // TODO -> This isn't used anywhere yet, but we do have a hard dependency, and need to check for these in the future.
+    // OHIF-229
     {
       id: 'org.ohif.cornerstone',
       version: '3.0.0',
