+++ platform/ui/src/components/ViewportActionBar/ViewportActionBar.jsx
@@ -24,6 +24,7 @@ const ViewportActionBar = ({
   // It shouldn't care that a tracking mode or SR exists.
   // Things like the right/left buttons should be made into smaller
   // Components you can compose.
+  // OHIF-200 ticket.
 
   const {
     label,
